# Improve histogram matching performance on unsigned integer data (resume #6209)