--- a/benchmarks/benchmark_exposure.py
+++ b/benchmarks/benchmark_exposure.py
@@ -1,5 +1,7 @@
 # See "Writing benchmarks" in the asv docs for more information.
 # https://asv.readthedocs.io/en/latest/writing_benchmarks.html
+import math
+
 import numpy as np
 
 from skimage import data, img_as_float
@@ -37,3 +39,55 @@ class ExposureSuite:
     def time_gamma_adjust_u8(self):
         for i in range(10):
             _ = exposure.adjust_gamma(self.image_u8)
+
+
+class MatchHistogramsSuite:
+
+    param_names = ["shape", "dtype", "multichannel"]
+    params = [
+        ((64, 64), (256, 256), (1024, 1024)),
+        (np.uint8, np.float32, np.float64),
+        (False, True),
+    ]
+
+    def _tile_to_shape(self, image, shape, multichannel):
+        n_tile = tuple(math.ceil(s / n) for s, n in zip(shape, image.shape))
+        if multichannel:
+            image = image[..., np.newaxis]
+            n_tile = n_tile + (3,)
+        image = np.tile(image, n_tile)
+        sl = tuple(slice(s) for s in shape)
+        return image[sl]
+
+    """Benchmark for exposure routines in scikit-image."""
+    def setup(self, shape, dtype, multichannel):
+        self.image = data.moon().astype(dtype, copy=False)
+        self.reference = data.camera().astype(dtype, copy=False)
+
+        self.image = self._tile_to_shape(self.image, shape, multichannel)
+        self.reference = self._tile_to_shape(self.reference, shape,
+                                             multichannel)
+        channel_axis = -1 if multichannel else None
+        self.kwargs = {'channel_axis': channel_axis}
+
+    def time_match_histogram(self, *args):
+        exposure.match_histograms(self.image, self.reference, **self.kwargs)
+
+    def peakmem_reference(self, *args):
+        """Provide reference for memory measurement with empty benchmark.
+
+        Peakmem benchmarks measure the maximum amount of RAM used by a
+        function. However, this maximum also includes the memory used
+        during the setup routine (as of asv 0.2.1; see [1]_).
+        Measuring an empty peakmem function might allow us to disambiguate
+        between the memory used by setup and the memory used by target (see
+        other ``peakmem_`` functions below).
+
+        References
+        ----------
+        .. [1]: https://asv.readthedocs.io/en/stable/writing_benchmarks.html#peak-memory
+        """
+        pass
+
+    def peakmem_match_histogram(self, *args):
+        exposure.match_histograms(self.image, self.reference, **self.kwargs)