--- a/benchmarks/benchmark_exposure.py
+++ b/benchmarks/benchmark_exposure.py
@@ -46,7 +46,7 @@ class MatchHistogramsSuite:
     param_names = ["shape", "dtype", "multichannel"]
     params = [
         ((64, 64), (256, 256), (1024, 1024)),
-        (np.uint8, np.float32, np.float64),
+        (np.uint8, np.uint32, np.float32, np.float64),
         (False, True),
     ]