--- a/skimage/exposure/histogram_matching.py
+++ b/skimage/exposure/histogram_matching.py
@@ -8,17 +8,27 @@ def _match_cumulative_cdf(source, template):
     Return modified source array so that the cumulative density function of
     its values matches the cumulative density function of the template.
     """
-    src_values, src_unique_indices, src_counts = np.unique(source.ravel(),
+    if source.dtype.kind == 'u':
+        src_lookup = source.ravel()
+        src_counts = np.bincount(src_lookup)
+        tmpl_counts = np.bincount(template.ravel())
+        tmpl_values = np.arange(len(tmpl_counts))
+        
+        tmpl_values_exist = tmpl_counts > 0
+        tmpl_counts = tmpl_counts[tmpl_values_exist]
+        tmpl_values = tmpl_values[tmpl_values_exist]
+    else:
+        src_values, src_lookup, src_counts = np.unique(source.ravel(),
                                                            return_inverse=True,
                                                            return_counts=True)
-    tmpl_values, tmpl_counts = np.unique(template.ravel(), return_counts=True)
+        tmpl_values, tmpl_counts = np.unique(template.ravel(), return_counts=True)
 
     # calculate normalized quantiles for each array
     src_quantiles = np.cumsum(src_counts) / source.size
     tmpl_quantiles = np.cumsum(tmpl_counts) / template.size
 
     interp_a_values = np.interp(src_quantiles, tmpl_quantiles, tmpl_values)
-    return interp_a_values[src_unique_indices].reshape(source.shape)
+    return interp_a_values[src_lookup].reshape(source.shape)
 
 
 @utils.channel_as_last_axis(channel_arg_positions=(0, 1))