## Description

This PR adds the suggestions from #6209. Benchmark results show a large improvement for unsigned integer types as compared to floating point types. The improvement is > 10x at larger sizes.

```

[100.00%] ··· benchmark_exposure.MatchHistogramsSuite.time_match_histogram 
[100.00%] ··· ============== =============== ============= =============
              --                                     multichannel       
              ------------------------------ ---------------------------
                  shape           dtype          False          True    
              ============== =============== ============= =============
                 (64, 64)      numpy.uint8     54.8±0.4μs     149±2μs   
                 (64, 64)      numpy.uint32    56.0±0.9μs     154±1μs   
                 (64, 64)     numpy.float32     287±1μs       843±7μs   
                 (64, 64)     numpy.float64    285±0.8μs      827±20μs  
                (256, 256)     numpy.uint8      354±7μs     1.30±0.02ms 
                (256, 256)     numpy.uint32     411±5μs     1.64±0.01ms 
                (256, 256)    numpy.float32   4.25±0.03ms   13.2±0.04ms 
                (256, 256)    numpy.float64    4.40±0.1ms    13.4±0.1ms 
               (1024, 1024)    numpy.uint8    5.33±0.04ms    23.2±0.5ms 
               (1024, 1024)    numpy.uint32   6.35±0.05ms    31.3±0.6ms 
               (1024, 1024)   numpy.float32    87.0±0.2ms     268±3ms   
               (1024, 1024)   numpy.float64    96.2±0.8ms     288±7ms   
              ============== =============== ============= =============
```


## Checklist

<!-- It's fine to submit PRs which are a work in progress! -->
<!-- But before they are merged, all PRs should provide: -->
- [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- Gallery example in `./doc/examples` (new features only)
- Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- Unit tests
- Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)
- Descriptive commit messages (see below)

<!-- For detailed information on these and other aspects see -->
<!-- the scikit-image contribution guidelines. -->
<!-- https://scikit-image.org/docs/dev/contribute.html -->

## For reviewers

<!-- Don't remove the checklist below. -->
- Check that the PR title is short, concise, and will make sense 1 year
  later.
- Check that new functions are imported in corresponding `__init__.py`.
- Check that new features, API changes, and deprecations are mentioned in
  `doc/release/release_dev.rst`.
- There is a bot to help automate backporting a PR to an older branch. For
  example, to backport to v0.19.x after merging, add the following in a PR
  comment: `@meeseeksdev backport to v0.19.x`
- To run benchmarks on a PR, add the `run-benchmark` label. To rerun, the label
  can be removed and then added again. The benchmark output can be checked in
  the "Actions" tab.