@pep8speaks 2022-04-24
Hello @grlee77! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`benchmarks/benchmark_exposure.py`](https://github.com/scikit-image/scikit-image/blob/e87595850061a463880ad1b85d7072d20fc32450/benchmarks/benchmark_exposure.py):

> [Line 88:80](https://github.com/scikit-image/scikit-image/blob/e87595850061a463880ad1b85d7072d20fc32450/benchmarks/benchmark_exposure.py#L88): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (88 > 79 characters)